Fixed #1 Slash Commandの3秒タイムアウト対応
