+++ .gitignore
@@ -1,5 +1,7 @@
 # Customize
 scripts/
+site-packages/
+packaged.yaml
 
 # Created by https://www.gitignore.io/api/osx,linux,python,windows,pycharm,visualstudiocode
