+++ README.md
@@ -1,1 +1,12 @@
 # Slack /ch Command
+
+A Slack slash command that searches for channels from channel name, purpose, topic.
+
+![screenshot](docs/images/screenshot.png)
+
+## Usage
+`/ch search KEYWORD`  
+`/ch list PREFIX`
+
+## License
+MIT
